User Interface update/bugfix
Added feature to highlight the active group in the sidebar. When you
load a group from an external page, you now know (finally!) exactly what
group you are in.

Also updated the notifications function and fixed some bugs with it.
There is now a 'clear all' button, to mark all notifications as 'seen'.